requirements.txt
asgiref==3.8.0
attrs==23.2.0
cfgv==3.4.0
chardet==5.2.0
distlib==0.3.8
Django==5.0.3
django-cors-headers==4.3.1
django-filter==24.1
django-sms==0.7.0
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
filelock==3.13.1
gunicorn==21.2.0
identify==2.5.35
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
nodeenv==1.8.0
packaging==24.0
pillow==10.2.0
platformdirs==4.2.0
pre-commit==3.6.2
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
reportlab==4.1.0
rpds-py==0.18.0
setuptools==69.2.0
sqlparse==0.4.4
tzdata==2024.1
uritemplate==4.1.1
virtualenv==20.25.1
